rename: 커스텀 훅 디렉토리 이름 변경
