+++ frontend/src/pages/my/MyPage.tsx
@@ -1,5 +1,4 @@
 import ProfileCard from './_components/profileCard'
-// import { type Profile, type ResponseChannelDto, type Stats } from '../../types/profile'
 import StatsCard from './_components/statsCard'
 import Targetbox from './_components/targetbox'
 import Conceptbox from './_components/conceptbox'
@@ -12,7 +11,7 @@ import Videos from '../../assets/icons/videos.svg?react'
 import Comment from '../../assets/icons/comment.svg?react'
 import { formatKoreanNumber } from '../../utils/format'
 import { useParams } from 'react-router-dom'
-import { useGetChannel } from '../../hooks/queries/useGetChannel'
+import { useGetChannel } from '../../hooks/my/useGetChannel'
 import { mapResponseToProfile } from '../../mappers/profile.ts/mapResponseToProfile'
 import { mapResponseStatCard } from '../../mappers/profile.ts/mapResponsetoStatCard'
 import { useEffect } from 'react'
