+++ frontend/src/pages/my/_components/videolist.tsx
@@ -2,8 +2,7 @@ import { useState } from 'react'
 import MyVideoCard from './myVideoCard'
 import MyShortsCard from './myShortsCard'
 import Pagination from '../../../components/Pagination'
-import { shortsData, videosData } from '../dummy'
-import { useGetChannelVideo } from '../../../hooks/queries/useGetChannelVideo'
+import { useGetChannelVideo } from '../../../hooks/my/useGetChannelVideo'
 import { useParams } from 'react-router-dom'
 import { mapResponseToVideoList } from '../../../mappers/profile.ts/mapResponseToVideo'
